Applied patch from Bill Sommerfeld at Apollo
The following patch to klist.c (relative to the beta-1 release)
improves the output formatting a bit.

a) Expired tickets (relative to the local clock) are reported as
EXPIRED, in large unfriendly letters.

b) Timestamp printing is collected into one procedure (better
modularity)

c) if the client name is the same as the default name, it's not
printed.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@2191 dc483132-0cff-0310-8789-dd5450dbe970