* in_tkt_ktb.c (keytab_keyproc): Do not check to see that the enctype of the key is identical; there are several equivalent DES enctypes.
* in_tkt_ktb.c (krb5_get_in_tkt_with_keytab): Removed the fancy
logic to only request the keytypes that correspond to those in
the keytab.  There were too many fencepost conditions that could
get you into trouble.  Either it should be there and *fully*
functional, or not in there at all.  Besides, there are too many
other components in Kerberos that expect the end-service to know
all its keys that this sanity check is overkill.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@7720 dc483132-0cff-0310-8789-dd5450dbe970