darwin build can't compile util/et test program
The autoconf tests correctly determine that sys_nerr exists, and that
no declaration is needed.  The test_et.c code checks the wrong condition,
though, and the system header declaration of sys_nerr conflicts with the
one in that source file.  (The native one is const.)

* test_et.c: Conditionalize sys_nerr declaration on NEED_SYS_ERRLIST, not
HAVE_SYS_ERRLIST.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15670 dc483132-0cff-0310-8789-dd5450dbe970