work around Solaris 9 pty-close bug
Create a LD_PRELOAD object, exitsleep, that will sleep for a short
time prior to calling the real exit() function.  This attempts to work
around a Solaris 9 kernel bug where output will get lost if it is
written to a pty immediately prior to the pty close.

ticket: new
component: krb5-build

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15864 dc483132-0cff-0310-8789-dd5450dbe970