A near complete re-write of the gss sample client on windows.  Supports the current protocol implemented in the Unix gss sample applications as well as a new User Interface making this one neat testing tool.
There are still many little kinks to get out in a future version.  The sliders
for the Call Count and the Message Count do not have text strings indicating
their current value.  They slide from 1 to 20.  And the known Mechanism
strings should be accessible in the drop down list.

A documentation file on how to use the tool would be a good addition.

ticket: 2144
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15987 dc483132-0cff-0310-8789-dd5450dbe970