As discussed on the krbdev mailing list, krb5_get_init_creds_password() suffered from a behavior in which it would unintentionally query a master KDC twice if in fact the KDC queried when krb5int_sendto() was called with use_master = 0 was in fact the master.  This resulted in more than an additional protocol operation.  There were two negative side effects. First, in the case of an incorrect password there would be two counts against the max retry attempts.  Second, in the case of hardware pre-auth and an expired password, the user would be asked to enter their expired password twice before being told it was expired.
This has been fixed by changing the use_master parameter into an in/out
parameter and modifying krb5int_sendto() to indicate which KDC it received
the response from.  This allows the use_master parameter to be set to
indicate whether or not the response came from a master KDC regardless
of whether a master KDC was requested.

ticket: new
target_version: next
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@16137 dc483132-0cff-0310-8789-dd5450dbe970