time-delta parser doesn't work on HP-UX 10
A build on Lenny's machine (built with gcc 2.7, thread support disabled, static
libraries only) is failing tests because the HP header files define "WS" as a
macro that expands to 2, while the delta-t grammar uses "WS" as the name of a
terminal symbol (which under bison turns into a C macro).

Renaming the terminal in the grammar (and custom lexer) fixes the problem.

* x-deltat.y (tok_WS): Renamed terminal from "WS", which conflicts with HP-UX
10 header files.  Updated all productions.
(mylex): Updated.
* Makefile.in ($(srcdir)/deltat.c): Enable dependencies in maintainer mode.
* deltat.c: Updated.

ticket: new
target_version: 1.4.1

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17110 dc483132-0cff-0310-8789-dd5450dbe970