Remove changebar.sty
The changebar.sty file found in the doc directory has questionable
licensing.  In particular, it is not the same changebar.sty found in
CTAN and has no well defined license specified in the file.  Since we
do not actually need changebar.sty, remove it.

ticket: new
Target_Version: 1.4.3
Tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17363 dc483132-0cff-0310-8789-dd5450dbe970