Network Identity Manager - Fix module loading when en_US locale cannot be loaded
The identity manager is designed for internationalization. However, it only ships with 
modules for the en_US locale.   Designing modules for other locales was beyond our the
reach of available resources.  This patch will force the use of en_US when modules 
matching the installed user and system locales cannot be found.

ticket: new 
status: resolved
target_version: 1.4.4
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17555 dc483132-0cff-0310-8789-dd5450dbe970