Move msg.c and marshall.c to common library
Add dllmain.c which contains the Windows DllMain() entry 
point.  This must be moved later to a Windows specific 
directory.

Cleanup type usage and function name references.  

All that is missing now from the client DLL is a definition
of cci_perform_rpc() which is the entry point to the IPC
routine.

git-svn-id: svn://anonsvn.mit.edu/krb5/branches/ccapi@18028 dc483132-0cff-0310-8789-dd5450dbe970