getaddrinfo code uses vars outside of storage duration
The calls to gethostbyname_r and friends were wrapped in macros that declared
automatic variables for the auxiliary storage needed, but the pointers returned
by the functions would be used outside of that scope, when the storage would no
longer be valid.

Changed the macro interfaces to define new types for the auxiliary storage, and
add a new argument to the macros referring to that storage.  Changed call sites
accordingly.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18180 dc483132-0cff-0310-8789-dd5450dbe970