allow multiple calls to krb5_get_error_message to retrieve message
(krb5int_get_error): Don't discard old message if the error codes don't match.
Try a little harder not to keep messages in the scratch buffer.  Return a copy
of the message, or "out of memory" in the scratch buffer.
(krb5int_vset_error): Try a little harder not to keep messages in the scratch
buffer.

ticket: new
target_version: 1.5

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18246 dc483132-0cff-0310-8789-dd5450dbe970