don't confuse profile iterator in 425 princ conversion
The profile iterator code hangs onto and uses the list of names passed
in.  The krb5_425_conv_principal code reuses that array when the
iterator may still be used.

* conv_princ.c (krb5_425_conv_principal): Use separate name arrays for
the iterator and the v4_realm lookup that may be done inside the
iteration loop.

ticket: new
target_version: 1.6
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@18748 dc483132-0cff-0310-8789-dd5450dbe970