Ensure consistancy between prototypes and functions
I am using an older compiler that is complaining that prototypes
do not match the functions they reference.  The issue is that a number
of prototypes are using "const int foo" while the function is "int foo".
From a caller sense it makes no difference - but the compiler is correct
they are different.

All is now consistant.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19169 dc483132-0cff-0310-8789-dd5450dbe970