NIM - Fix taskbar button visibility on Vista
Although the Platform SDK docs suggest using ITaskbarList to add/remove
a taskbar button on the fly, it doesn't work on Vista.  Instead we will
just set the window style to use WS_EX_APPWINDOW and be done with it.

No change in the NIM version number.

ticket: new
component: windows
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19440 dc483132-0cff-0310-8789-dd5450dbe970