don't check for readability resolving SRVTAB: keytab
Greg Hudson has pointed out that the FILE: and SRVTAB: keytab types differ in
when they report errors for files not existing.  The implementations of ANY:
floating around cope better if they consistently accept names of files that
aren't present, and report the error later when fetching information.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19563 dc483132-0cff-0310-8789-dd5450dbe970