Be more aggressive about forcing the new credentials dialog to the top.
Many applications call GSSAPI functions from the main GUI thread.
Doing so blocks the window message pump for that application.  When
the GSSAPI call results in the new credentials dialog being displayed,
it will be unable to bring itself to the top and grab the user input
focus because the foreground window is not responding.  This patch
makes the new credentials window a top-most window at creation time
that results in it being created at the top of the Z-order.  It can
then flash the window to indicate that it requires user attention.

ticket: 5584

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19629 dc483132-0cff-0310-8789-dd5450dbe970