Patch developed by kpkoch with style changes from jaltman
The size/position of the main application window is 
internally updated in response to WM_MOVE messages but is
only written to the registry after a timeout period.  This
is done due to the large number of WM_MOVE messages that 
can be delivered during a windows drag / resize operation
involving the user or explorer shell's tile and cascade 
operations.  (or those involving third party desktop managers.)

In NIM 1.8 two different application view modes (standard 
and advanced) replaced the single view mode in previous 
releases.  The size/position update logic was not modified
to take into consideration the possibility that a user might
move/resize the window and then quickly toggle modes before
the new location or size were recorded to the registry.

This change ensures that when a mode change occurs, via a
call to khm_set_main_window_mode(), that the current 
location/size will be written to the registry and any 
outstanding timer, MW_RESIZE_TIMER, will be cleared.

The logic to save the location/size has been extracted 
into the new static function main_wnd_save_sizepos().

main_wnd_save_sizepos() is only called after the application
window has been created.

ticket: 5613

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19760 dc483132-0cff-0310-8789-dd5450dbe970