need more dylib_file specs for darwin
Currently the KDB LDAP plugin won't build on Mac OS X 10.4 if a tree
hasn't been previously installed, because it can't find the libraries
that we haven't installed yet.  (Finding earlier versions isn't
sufficient, if symbols are needed that are not present in the
installed versions.)

Add -dylib_file specs for libkadm5srv and libkdb to LDCOMBINE, in
addition to libkrb5support that was already there.

Unfortunately, this makes shlib.conf dependent on more library version
numbers.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19771 dc483132-0cff-0310-8789-dd5450dbe970