NIM: resource leak in khui_action_trigger()
The khui_action_trigger() function in the Network Identity Manager API
creates a copy of the current user interface context (which contains
information about the credentials and identities that are currently
selected) before triggering the requested action.  Then it will use
the copy to restore the user interface context to its previous state.
A coding error results in the copy of the context to never be
released.  As a result there is a resource leak.

The attached patch fixes the leak by releasing the context once it
is no longer needed.

ticket: new
component: windows

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19852 dc483132-0cff-0310-8789-dd5450dbe970