NIM: Identity Configuration Panel Fixes
In the identity configuration panel of Network Identity Manager, the
user can specify a new identity which can then be configured.

The existing code didn't check if the identity specified by the user
already exists.  The patch adds the check.  If the identity already
exists, the user is notified as such.

Another bug prevented the user from configuring an identity that was
added back in following the deletion of the same identity during the
same session.  The deleted status of the identity was not reset when
it was added back.

Additionally, this patch adds code that has already been added to the
new credentials dialog to apply Windows XP theme textures to the child
dialogs used as tab panels in the configuration dialog.  Child dialogs
don't automatically adjust the theme settings based on whether it is
nested inside a tab control.  The theme must be applied manually.

ticket: new
component: windows

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19853 dc483132-0cff-0310-8789-dd5450dbe970