NIM: Handle WM_PAINT messages without update regions
It is possible to receive a WM_PAINT message in Windows without there
being an update region.  For example, this can be caused by someone
calling RedrawWindow() with the RDW_INTERNALPAINT flag set.  In this
case, GetUpdateRect() will indicate that there is no update region and
calling BeginPaint()/EndPaint() results in incorrect behavior.

The credentials window in Network Identity Manager needs to perform
special handling for this case by obtaining a proper device context
and completing the drawing operation.

ticket: new 
component: windows

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19859 dc483132-0cff-0310-8789-dd5450dbe970