NIM: External changes to default identity are improperly reflected by krb5 provider
The Kerberos v5 identity provider for Network Identity Manager
monitors the "Software\MIT\kerberos5" registry key for the logged in
user for changes to the "ccname" value.  If a change is noticed, it
would query the Kerberos v5 library for the default credentials cache
and attempt to determine the new default identity, which it would then
communicate to the Network Identity Manager application.

When the identity provider queried the Kerberos v5 library after a
registry change notification, it used a cached krb5_context for the
thread.  The default credentials cache found using this krb5_context
may not be what the registry specified.

This patch modifies the code in k5_ccname_monitor_thread() to create a
use a new krb5_context when querying for the default credentials cache
following a registry change notification.  Doing so ensures that
Kerberos v5 library takes the new registry value into account.

ticket: new 
component: windows

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@19866 dc483132-0cff-0310-8789-dd5450dbe970