NIM: BUG: APP: identity data corruption
The credentials display in Network Identity Manager uses a cached set
of properties for each identity that has credentials.  The existing
code used the wrong field of the khui_credwnd_identity structure to
store the credentials type name associated with an identity, resulting
in a corrupt data structure.

This patch fixes the code to use the correct field.

ticket: new 

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20030 dc483132-0cff-0310-8789-dd5450dbe970