NIM: BUG: APP: notification icon state
Currently, the notification icon that is created by Network Identity
Manager indicates the status of all the known credentials.  If any
credential belonging to any identity is expired or in a critical
state, the icon will change color to indicate this.

This patch changes the behavior to only indicate the status of
credentials belonging to the default identity.  The expiration state
of credentials that do not belong to the default identity no longer
affect the appearance of the notification icon.

ticket: new 

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20031 dc483132-0cff-0310-8789-dd5450dbe970