NIM: BUG: APP: notification icon tooltip wrong string
The Network Identity Manager notification icon can display a tooltip
when the user hovers the mouse cursor over it.  It is currently used
to indicate the default identity (if one is found).  However, when
retrieving the name of the default identity, the size of the buffer
was left unspecified.  This patch specifies the correct buffer size.

The function called with the incorrect buffer size was
kcdb_identity_get_name().  That function does not write more than
KCDB_IDENT_MAXCCH_NAME characters regardless of the size of the buffer
specified, and the buffer that was passed in is allocated to be this
size.  No buffer overrun was present in the existing code, although
the behavior was incorrect.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20034 dc483132-0cff-0310-8789-dd5450dbe970