The khm_show_main_window() function is no longer called at startup with khm_nCmdShow == SW_SHOWMINIMIZED in order to hide the main application by calling khm_hide_main_window(). Instead, the main application window is simply never shown.
As a result, khm_show_main_window() needs to respond to 
khm_nCmdShow == SW_SHOWMINIMIZED not by hiding the window
but by changing the khm_nCmdShow state to SW_SHOW and then
calling ShowWindow().

This change will address the problem whereby "Show NIM Window"
had to be triggered twice by the user when the process
was started in a minimized state.

ticket: 5842

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20164 dc483132-0cff-0310-8789-dd5450dbe970