signed vs unsigned char * warnings in kdb_xdr.c
load and store functions in k5-platform.h take an unsigned char *, whereas
kdb_xdr.c was using a char * for decoding.  This resulted in pages of warnings
in the parsing code.  Switched to using an unsigned char * and cast in the
couple places where a char * is needed.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20343 dc483132-0cff-0310-8789-dd5450dbe970