Assign fixed ordinals to comerr32.dll exports
All of the other libraries on Windows have fixed assignments
of ordinals to the exported functions.  Assign the ordinals 
that were in use in the last public release, kfw 3.2.2, so 
that they will remain constant into the future in case additional
exports are added to the library.

ticket: new
component: krb5-libs
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20553 dc483132-0cff-0310-8789-dd5450dbe970