The NIM error reporting functions (in src/windows/identity/kherr ) keep track of the the error message with the highest severity level that was reported for a specific error reporting context.  However, if another error message of the same severity is reported, the error message being tracked will be updated to be the newly received error.
The user will often only be notified of the error message that was
tracked for a specific operation.  Therefore, tracking the last message
with the highest priority has the unfortunate side-effect of not
reporting the cause of a failure.

This patch changes the condition for updating the tracked error message
to be the first message with the highest severity.

ticket: 6047
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20573 dc483132-0cff-0310-8789-dd5450dbe970