Integrate FAST in to client AS sending
Functions to generate  FAST request and to manage client fast state.  Integrate into client  AS req loop.

Most of this is stub code although the integration points and arguments should be correct.

* Call into fast to prepare the request body.  If FAST is being used,
this may end up hiding the names in the future in the outer request.

* Call into FAST to prepare the request before sending.  This will
generate the FAST padata  and return an encoded outer request.

* Call into fast to handle error replies, potentially extracting
padata and information on whether to continue processing.

git-svn-id: svn://anonsvn.mit.edu/krb5/branches/fast@22126 dc483132-0cff-0310-8789-dd5450dbe970