CVE-2009-0844 (1.6.x) SPNEGO can read beyond buffer end
pull up rxxxxx from trunk

SPNEGO can read beyond the end of a buffer if the claimed DER length
exceeds the number of bytes in the input buffer. This can lead to
crash or information disclosure.

Thanks to Apple for reporting this vulnerability and providing
patches.

ticket: 6446
tags: pullup
target_version: 1.6.4
version_fixd: 1.6.4

git-svn-id: svn://anonsvn.mit.edu/krb5/branches/krb5-1-6@22179 dc483132-0cff-0310-8789-dd5450dbe970