minor memory leak in 'none' replay cache type
The replay cache type implementations are responsible for freeing the
main rcache structure when the cache handle is closed.  The 'none'
rcache type wasn't doing this, resulting in a small memory leak each
time such a cache was opened and closed.  Not a big deal for a server
process servicing a single client, but it could accumulate (very very
slowly) for a long-running server.

ticket: 6514
tags: pullup
target_version: 1.7.1
version_reported: 1.7

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@22417 dc483132-0cff-0310-8789-dd5450dbe970