quoting bug causes solaris pre-10 thread handling bugs
Quoting problems in pattern matching on the OS name cause Solaris
versions up through 9 to not be properly recognized in the
thread-system configuration setup.  This causes our libraries to make
the erroneous assumption that valid thread support routines are
available on all Solaris systems, rather than just assuming it for
Solaris 10 and later.

The result is assertion failures like this one reported by Meraj
Mohammed and others:

  Assertion failed: k5int_i->did_run != 0, file krb5_libinit.c, line 63

Thanks to Tom Shaw for noticing the cause of the problem.

The bug may be present in the 1.6.x series as well.

ticket: 6579
target_version: 1.7.1
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@23144 dc483132-0cff-0310-8789-dd5450dbe970