Pullup to 1.7-branch is only for the test case, as krb5-1.7 behaved correctly for these checksums.
Fix regression in MD4-DES and MD5-DES keyed checksums.  The original
key was being used for the DES encryption, not the "xorkey".  (key
with each byte XORed with 0xf0)

Add a test case that will catch future regressions of this sort, by
including a verification of a "known-good" checksum (derived from a
known-to-be-interoperable version of the implementation).

ticket: 6584
target_version: 1.7.1
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@23361 dc483132-0cff-0310-8789-dd5450dbe970