Implement gss_set_neg_mechs
Implement gss_set_neg_mechs in SPNEGO by intersecting the provided
mech set with the mechanisms available in the union credential.  As
we now need space to hold the mech set, the SPNEGO credential is now
a structure and not just a mechglue credential.

t_spnego.c is a test program which exercises the new logic.  Like the
other GSSAPI tests, it is not run as part of "make check" at this
time.

ticket: 6658
target_version: 1.8
tags: pullup

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@23715 dc483132-0cff-0310-8789-dd5450dbe970