Require IPv6 support
The configure-time options to enable and disable IPv6 support have
been deprecated for some time, but the checks for OS support were
kept.  This removes those checks, and unconditionally compiles in the
IPv6 support.

There was a configure-time test to see if the macro INET6 needed to be
defined in order to enable (visibility of) OS support for IPv6, which
was needed on an IRIX system we tested with.  That check is retained,
but the revised code is untested on IRIX.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25719 dc483132-0cff-0310-8789-dd5450dbe970