Only list --enable-dns-for-realm once
If we list something as "commonly used", we seem to not also
list it with the rest of the options.  This has the advantage
of not requiring us to remember to update two things for future
changes, but the disadvantage of requiring users to look in two
places for options.  Stick with the prevailing form for now.