Use '---' for em dashes in rst source
Our sphinx configuration uses SmartyPants, which produces smart
quotes and dashes in HTML output, using '--' for en dash and
'---' for em dash.  (This is also the LaTeX convention.)
These points in the text are meant to be em dashes, so format them
as such.  Also standardize on no spaces around the dash per
Chicago Manual of Style (and others).