Document absolute times for 'kinit -s'
kinit's -s argument takes a parameter, which is first interpreted
as a time interval ("deltat"), in the same format used for
the -l and -r arguments.  However, if that fails, the
time argument is interpreted as an absolute time, in one of
twelve different possible formats, some of which are subject to
localization via strptime().  Explicitly listing all twelve
possible interpretations is overkill, but give the user a hint that
absolute times are possible at all.