Remove dash from man page rst source
This page gets rendered for the web with Sphinx but is also turned
into the krb5_conf.5 manual page.  We need to use three-hyphen
em dashes for the Sphynx config, but those are a bit long for
monospace terminal output.  Since the dash here can easily be
changed to a comma, do so, and avoid the conflict of formatting.