Try harder not to use clock_gettime in verto-k5ev
We already define EV_USE_REALTIME to 0 to avoid the use of
clock_gettime() (to avoid depending on librt).  But in some build
environments libev can detect support for a monotonic clock, which
also results in using clock_gettime().  Define EV_USE_MONOTONIC to 0
as well to prevent this.

ticket: 7350 (new)
target_version: 1.10.4
tags: pullup